refactored contexts to be part of model
