+++ packages/livedoc-mocha/_src/app/index.ts
@@ -6,8 +6,8 @@ import * as mochaTest from "mocha/lib/test";
 import * as model from "./model";
 import { LiveDoc } from "./livedoc";
 import { LiveDocContext } from "./LiveDocContext";
-import { BddContext } from "./BddContext";
-import { ScenarioOutlineContext } from "./ScenarioOutlineContext";
+import { BddContext } from "./model/BddContext";
+import { ScenarioOutlineContext } from "./model/ScenarioOutlineContext";
 import { LiveDocGrammarParser } from "./parser/Parser";
 import { LiveDocDescribe } from "./model/LiveDocDescribe";
 import { RuleViolations } from "./model/RuleViolations";
